Simple Ftp Server And Client Program Using Java
I developed this project for the lesson Computer Networks at the
University of Fatih Sultan Mehmet.

Developed with Java, to use this program, first make active the FTP
server.

In order to start the server, you must define the server name and port
number. Then, the client can get access to a designated folder on the
server side. If you are not registered if the user allows the user to
create, and then you can view the files in your own folder with your
user name.

You can look at the screenshots if you wish. For more detailed
information burhangok@yahoo.com you can mail it to the address.